some steps into a conditional layout
